Harden the code and crawl multireddits
Code will now no longer throw when reddit’s API returns something unexpected (like HTML instead of JSON).

Methodology is updated. Instead of trying to search for all occurences of a phrase of all time, it narrows the search to specific programming-related subreddits, and it goes month by month. This makes it almost sure that we get all the articles, and that there are no false positives.